feat(core): 支持/lib下的jar

+++ core/src/main/java/org/springplugin/core/context/SpringNamedContext.java
@@ -112,9 +112,12 @@ public class SpringNamedContext extends AbstractPluginContext implements PluginC
                     log.error(String.format("plugin classloader close fail, %s", name), e);
                 }
             }
-            if (!FileUtils.deleteQuietly(new File(SpringPluginClassLoader.LOAD_PATH + name))) {
-                log.error("unload plugin fail, can't delete plugin file: {}", name);
-                return false;
+            final SpringPluginProperties.Debug debug = pluginProperties.getDebug();
+            if (!debug.isResourceCache()) {
+                if (!FileUtils.deleteQuietly(new File(SpringPluginClassLoader.LOAD_PATH + name))) {
+                    log.error("unload plugin fail, can't delete plugin file: {}", name);
+                    success = false;
+                }
             }
         }
         return true;
